+++ resources/container.rb
@@ -21,7 +21,7 @@ attribute :entrypoint, :kind_of => [String]
 attribute :env, :kind_of => [String, Array]
 attribute :env_file, :kind_of => [String]
 attribute :expose, :kind_of => [Fixnum, String, Array]
-attribute :force, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :force, :kind_of => [TrueClass, FalseClass]
 attribute :hostname, :kind_of => [String]
 attribute :id, :kind_of => [String]
 attribute :init_type, :kind_of => [FalseClass, String], :default => node['docker']['container_init_type']
@@ -38,8 +38,8 @@ attribute :port, :kind_of => [Fixnum, String, Array]
 attribute :privileged, :kind_of => [TrueClass, FalseClass]
 # public_port attribute is DEPRECATED
 attribute :public_port, :kind_of => [Fixnum]
-attribute :publish_exposed_ports, :kind_of => [TrueClass, FalseClass], :default => false
-attribute :remove_automatically, :kind_of => [TrueClass, FalseClass], :default => false
+attribute :publish_exposed_ports, :kind_of => [TrueClass, FalseClass]
+attribute :remove_automatically, :kind_of => [TrueClass, FalseClass]
 attribute :repository, :kind_of => [String]
 attribute :run, :kind_of => [String]
 attribute :signal, :kind_of => [String]